move and improve command add_fake_6002x_students
